+++ modules/engine/lib/engine/describe-route.js
@@ -48,9 +48,7 @@ var arr = [], routes = opts.routes, context = opts.context, record, otherVerbs =
     record = routes.simpleMap[statement.method + ':' + statement.path.value];
 
     if(!record) {
-        cb({
-            message: 'No such route ' + statement.path.value + ' for HTTP method "' + 'statement.method"'
-        });
+        cb('No such route ' + statement.path.value + ' for HTTP method "' + 'statement.method"');
     }
 
     cb(null, {