+++ modules/engine/lib/engine/insert.js
@@ -60,14 +60,10 @@ exports.exec = function(opts, statement, cb, parentEvent) {
         // Get the resource
         table = opts.tempResources[name] || tables[name];
         if(!table) {
-            return insertTx.cb({
-                message: 'No such resource ' + name
-            });
+            return insertTx.cb('No such table ' + name);
         }
         if(!table.insert) {
-            return insertTx.cb({
-                message: 'Table ' + statement.source.name + ' does not support insert'
-            });
+            return insertTx.cb('Table ' + statement.source.name + ' does not support insert');
         }
 
         httpRequest.exec({